+++ .gitignore
@@ -1,4 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+.pnpm-store
 
 # dependencies
 node_modules
