+++ apps/api/src/lib/env.ts
@@ -25,6 +25,10 @@ const envSchema = z.object({
 
   TWITTER_CLIENT_ID: z.string().default(''),
   TWITTER_CLIENT_SECRET: z.string().default(''),
+
+  SENDGRID_API_KEY: z.string().default(''),
+  SENDGRID_SENDER_EMAIL: z.string().default(''),
+  SENDGRID_SENDER_NAME: z.string().default(''),
 })
 
 export const env = new Proxy({} as z.infer<typeof envSchema>, {
