+++ apps/api/src/routes/status.test.ts
@@ -1,39 +1,66 @@
-import type { FastifyInstance } from 'fastify'
-import { afterAll, beforeAll, describe, expect, test } from 'vitest'
+import { describe, expect, test } from 'vitest'
 import { createServer } from '../server.js'
 
-describe('server', () => {
-  let testServer: FastifyInstance
+describe('Status Routes', () => {
+  test('GET /api/status - should return system health status', async () => {
+    const app = createServer()
 
-  beforeAll(async () => {
-    const server = await createServer({ logger: false })
-    if (!server) {
-      throw new Error('Server is null')
-    }
-    testServer = server
-    await testServer.ready()
-  })
-
-  afterAll(async () => {
-    if (!testServer) {
-      return
-    }
-
-    await testServer.close()
-  })
-
-  test('status endpoint returns 200', async () => {
-    const response = await testServer.inject({
-      method: 'GET',
-      url: '/status',
-    })
-    const body = JSON.parse(response.body) as {
-      up: boolean
-      isAuth: boolean
+    const res = await app.request('/api/status')
+    const body = (await res.json()) as {
+      status: string
       serverTime: string
       uptime: number
+      database: string
     }
-    expect(response.statusCode).toBe(200)
-    expect(body.up).toEqual(true)
+
+    expect(res.status).toBe(200)
+    expect(body).toMatchObject({
+      status: 'ok',
+      serverTime: expect.any(String),
+      uptime: expect.any(Number),
+      database: 'connected',
+    })
+
+    // Verify serverTime is a valid ISO string
+    expect(() => new Date(body.serverTime)).not.toThrow()
+
+    // Verify uptime is positive
+    expect(body.uptime).toBeGreaterThan(0)
+  })
+
+  test('GET /api/status/health - should return simple health check', async () => {
+    const app = createServer()
+
+    const res = await app.request('/api/status/health')
+    const body = (await res.json()) as {
+      status: string
+      serverTime: string
+    }
+
+    expect(res.status).toBe(200)
+    expect(body).toMatchObject({
+      status: 'ok',
+      serverTime: expect.any(String),
+    })
+
+    // Verify serverTime is a valid ISO string
+    expect(() => new Date(body.serverTime)).not.toThrow()
+  })
+
+  test('GET /api/status - should handle database connection errors gracefully', async () => {
+    // This test would require mocking the database to fail
+    // For now, we'll just test the happy path since database mocking
+    // in integration tests can be complex
+    const app = createServer()
+
+    const res = await app.request('/api/status')
+
+    // Should either succeed (200) or fail gracefully (500)
+    expect([200, 500]).toContain(res.status)
+
+    const body = (await res.json()) as Record<string, unknown>
+    expect(body).toHaveProperty('status')
+    expect(body).toHaveProperty('serverTime')
+    expect(body).toHaveProperty('uptime')
   })
 })
