+++ apps/api/src/routes/status.ts
@@ -1,20 +1,37 @@
-import type { FastifyPluginAsync } from 'fastify'
-import fp from 'fastify-plugin'
+import { db } from '@hominem/utils/db'
+import { health } from '@hominem/utils/schema'
+import { Hono } from 'hono'
 
-// Status/health endpoint
-const statusPlugin: FastifyPluginAsync = async (server) => {
-  // Simple health check that doesn't require authentication
-  server.get('/health', async () => {
-    return { status: 'ok', serverTime: new Date().toISOString() }
-  })
+export const statusRoutes = new Hono()
 
-  server.get('/status', async (request) => {
-    const isAuth = !!request.userId
-    const serverTime = new Date().toISOString()
-    const uptime = process.uptime()
+// System health check endpoint
+statusRoutes.get('/', async (c) => {
+  try {
+    // Test database connection
+    await db.select().from(health).limit(1)
 
-    return { up: true, isAuth, serverTime, uptime }
-  })
-}
+    return c.json({
+      status: 'ok',
+      serverTime: new Date().toISOString(),
+      uptime: process.uptime(),
+      database: 'connected',
+    })
+  } catch (error) {
+    console.error('Health check failed:', error)
+    return c.json(
+      {
+        status: 'error',
+        serverTime: new Date().toISOString(),
+        uptime: process.uptime(),
+        database: 'disconnected',
+        error: error instanceof Error ? error.message : 'Unknown error',
+      },
+      500
+    )
+  }
+})
 
-export default fp(statusPlugin)
+// Legacy /health endpoint for backward compatibility
+statusRoutes.get('/health', async (c) => {
+  return c.json({ status: 'ok', serverTime: new Date().toISOString() })
+})
