+++ apps/api/test/api-test-utils.ts
@@ -1,5 +1,6 @@
-import type { FastifyInstance } from 'fastify'
+import type { Hono } from 'hono'
 import { afterAll, beforeAll, beforeEach, expect, vi } from 'vitest'
+import type { AppEnv } from '../src/server.js'
 import { createServer } from '../src/server.js'
 
 /**
@@ -13,18 +14,18 @@ export const globalMocks = {
   },
 
   // Auth middleware mock
-  verifyAuth: vi.fn((request, reply, done) => {
-    request.userId = 'test-user-id'
-    done()
+  verifyAuth: vi.fn((c, next) => {
+    c.set('userId', 'test-user-id')
+    return next()
   }),
 
   // Rate limit middleware mocks
-  rateLimit: vi.fn((request, reply, done) => {
-    done()
+  rateLimit: vi.fn((c, next) => {
+    return next()
   }),
 
-  rateLimitImport: vi.fn((request, reply, done) => {
-    done()
+  rateLimitImport: vi.fn((c, next) => {
+    return next()
   }),
 }
 
@@ -32,18 +33,12 @@ export const globalMocks = {
  * Creates a test server instance with common setup
  */
 export const createTestServer = async (options: { logger?: boolean } = {}) => {
-  const server = await createServer({ logger: options.logger ?? false })
+  const server = createServer()
   if (!server) {
     throw new Error('Server is null')
   }
 
-  // Setup mock queues
-  server.queues = {
-    plaidSync: globalMocks.queue,
-    importTransactions: globalMocks.queue,
-  } as unknown as typeof server.queues
-
-  await server.ready()
+  // Hono doesn't need explicit setup like Fastify
   return server
 }
 
@@ -51,7 +46,7 @@ export const createTestServer = async (options: { logger?: boolean } = {}) => {
  * Common test lifecycle hooks for API route tests
  */
 export const useApiTestLifecycle = () => {
-  let testServer: FastifyInstance
+  let testServer: Hono<AppEnv>
 
   beforeAll(async () => {
     testServer = await createTestServer()
@@ -62,8 +57,9 @@ export const useApiTestLifecycle = () => {
   })
 
   afterAll(async () => {
+    // Hono doesn't need explicit cleanup like Fastify
     if (testServer) {
-      await testServer.close()
+      // No close method needed for Hono
     }
   })
 
@@ -87,7 +83,7 @@ export interface ApiResponse {
  * Helper for making authenticated requests
  */
 export const makeAuthenticatedRequest = async (
-  server: FastifyInstance,
+  server: Hono<AppEnv>,
   options: {
     method: 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH'
     url: string
@@ -99,42 +95,41 @@ export const makeAuthenticatedRequest = async (
     'x-user-id': 'test-user-id',
     ...options.headers,
   }
-  if (options.payload) {
-    headers['Content-Type'] = 'application/json'
+
+  const requestInit: RequestInit = {
+    method: options.method,
+    headers,
   }
 
-  return server.inject({
-    method: options.method,
-    url: options.url,
-    payload: options.payload,
-    headers,
-  })
+  if (options.payload) {
+    headers['Content-Type'] = 'application/json'
+    requestInit.body = JSON.stringify(options.payload)
+  }
+
+  const request = new Request(`http://localhost${options.url}`, requestInit)
+  return server.fetch(request)
 }
 
 /**
  * Helper for parsing JSON responses with proper typing
  */
-export const parseJsonResponse = <T = ApiResponse>(response: { body: string }): T => {
-  return JSON.parse(response.body) as T
+export const parseJsonResponse = async <T = ApiResponse>(response: Response): Promise<T> => {
+  return (await response.json()) as T
 }
 
 /**
  * Common assertion helpers
  */
-export const assertSuccessResponse = <T>(response: { statusCode: number; body: string }) => {
-  const body = parseJsonResponse<T>(response)
-  expect(response.statusCode).toBe(200)
-  // Don't require specific success field - let API consumers handle the response structure
+export const assertSuccessResponse = async <T>(response: Response) => {
+  const body = await parseJsonResponse<T>(response)
+  expect(response.status).toBe(200)
   return body
 }
 
-export const assertErrorResponse = (
-  response: { statusCode: number; body: string },
-  expectedStatus = 500
-) => {
-  const body = parseJsonResponse(response)
-  expect(response.statusCode).toBe(expectedStatus)
-  // Focus on HTTP status codes rather than specific response structure
+export const assertErrorResponse = async (response: Response, expectedStatus = 500) => {
+  const body = await parseJsonResponse(response)
+  expect(response.status).toBe(expectedStatus)
+
   if (expectedStatus >= 400) {
     // Allow various error response formats - either { success: false } or { error: "message" }
     const hasErrorIndicator =
