--- apps/cli/drizzle.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from 'drizzle-kit'
-import { env } from './src/env'
-
-export default defineConfig({
-  out: './src/db/migrations',
-  schema: './src/db/schema.ts',
-  dialect: 'sqlite',
-  dbCredentials: {
-    url: env.DB_PATH,
-  },
-})
