+++ docker-compose.yml
@@ -23,6 +23,29 @@ services:
       timeout: 5s
       retries: 5
 
+  postgres-test:
+    build:
+      context: .
+      dockerfile: docker/postgres.dockerfile
+    container_name: hominem-postgres-test
+    restart: always
+    environment:
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_DB: hominem-test
+    networks:
+      - hominem
+    ports:
+      - 4433:5432
+    volumes:
+      - hominem-postgres-test:/var/lib/postgresql/data
+      - ./docker/init-extensions.sql:/docker-entrypoint-initdb.d/init-extensions.sql
+    healthcheck:
+      test: ["CMD-SHELL", "pg_isready -U postgres"]
+      interval: 10s
+      timeout: 5s
+      retries: 5
+
   rabbitmq:
     image: rabbitmq:3-management
     container_name: hominem-rabbitmq
@@ -52,4 +75,5 @@ networks:
 
 volumes:
   hominem-postgres:
+  hominem-postgres-test:
   hominem-redis:
